Update io_fews.py
FutureWarning: Could not cast to float64, falling back to object. This behavior is deprecated. In a future version, when a dtype is passed to 'DataFrame', either all columns will be cast to that dtype, or a TypeError will be raised.
  ts = pd.DataFrame(events, index=index, dtype=float)

+++ hydropandas/io/io_fews.py
@@ -394,7 +394,8 @@ def read_xml_root(
             index = pd.to_datetime(
                 [d + " " + t for d, t in zip(date, time)], errors="coerce"
             )
-            ts = pd.DataFrame(events, index=index, dtype=float)
+            ts = pd.DataFrame(events, index=index)
+            ts.loc[:, "value"] = ts.loc[:, "value"].astype(float)
 
             if remove_nan and (not ts.empty):
                 ts.dropna(subset=["value"], inplace=True)
